Former-commit-id: 9fa013c8e7abaaddbed801be42f87633ec968eca [formerly c064351aefd5789094a239c66cdb611ba2b55375] [formerly 918ff154e1bd995c3374b2bd49bae015bdcd1b4a [formerly 0c11f99805b786370234f05cbd3da6c3efd69337]] [formerly 796380dd53e8a3a232f3e9e4f6b1b4b599f6597f] [formerly 9b44ec25d17d998243dc38d9d1c3d9fb75fd7b5f [formerly 9b44ec25d17d998243dc38d9d1c3d9fb75fd7b5f [formerly adfeb1306751695abe2eaaf23d62ed4f3baa2430]] [formerly 796380dd53e8a3a232f3e9e4f6b1b4b599f6597f [formerly 796380dd53e8a3a232f3e9e4f6b1b4b599f6597f [formerly 56dc338129ed46612df4db1926e0b7e91551b60d]]]]
Former-commit-id: 9b44ec25d17d998243dc38d9d1c3d9fb75fd7b5f [formerly 9b44ec25d17d998243dc38d9d1c3d9fb75fd7b5f [formerly adfeb1306751695abe2eaaf23d62ed4f3baa2430]]
Former-commit-id: 41055f607367027477ecc3be60e146bbfbffd00f [formerly bef1582f44a5bdfb56ef35bab2cb2b20aa900312] [formerly 5e62debf5b4d24505bedfe384debe4830e470577 [formerly 12c854e7cc66f4fd55c067ee414a3797e0cac5d2]]
Former-commit-id: 0844806ad426ed29171b8a70ce80689befafcc7a [formerly 60f9feed3efc51b1179e84a664cd1775442ffe07]
Former-commit-id: dfd97c84821a1335becc83ef79970a6c071d973b

+++ cave/com.raytheon.uf.viz.coopprecip/src/com/raytheon/uf/viz/coopprecip/CoopPrecipPlotInfoRetriever.java
@@ -27,8 +27,8 @@ import java.util.List;
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 
-import com.raytheon.uf.common.inventory.exception.DataCubeException;
 import com.raytheon.uf.common.dataquery.requests.RequestConstraint;
+import com.raytheon.uf.common.inventory.exception.DataCubeException;
 import com.raytheon.uf.common.pointdata.PointDataContainer;
 import com.raytheon.uf.common.pointdata.PointDataView;
 import com.raytheon.uf.common.time.DataTime;
@@ -50,6 +50,7 @@ import com.raytheon.viz.pointdata.rsc.retrieve.AbstractPlotInfoRetriever;
  * Date         Ticket#    Engineer    Description
  * ------------ ---------- ----------- --------------------------
  * Oct 9, 2009            bsteffen     Initial creation
+ * Aug 08, 2014 3477      bclement     changed plot info locations to floats
  * 
  * </pre>
  * 
@@ -75,9 +76,8 @@ public class CoopPrecipPlotInfoRetriever extends AbstractPlotInfoRetriever {
             for (int uriCounter = 0; uriCounter < pdc.getCurrentSz(); uriCounter++) {
                 PointDataView pdv = pdc.readRandom(uriCounter);
                 PlotInfo stationInfo = new PlotInfo();
-                stationInfo.latitude = pdv.getNumber("latitude").doubleValue();
-                stationInfo.longitude = pdv.getNumber("longitude")
-                        .doubleValue();
+                stationInfo.latitude = pdv.getNumber("latitude").floatValue();
+                stationInfo.longitude = pdv.getNumber("longitude").floatValue();
                 stationInfo.dataTime = new DataTime(new Time(
                         pdv.getLong("time")));
                 stationInfo.stationId = pdv.getString("stationId");
